Webs

+++ app/Http/Controllers/Api/V1/Users/OfferController.php
@@ -9,6 +9,7 @@ use App\Models\Supplier;
 use App\Services\MobileNotificationServices;
 use App\Traits\Images;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
@@ -21,10 +22,8 @@ class OfferController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Offer::class);
         $offers = Offer::all();
-        foreach ($offers as $offer) {
-            $offer->image = asset("storage/$offer->image");
-        }
         return response()->json($offers, 200);
     }
 
@@ -36,7 +35,7 @@ class OfferController extends Controller
     public function show(string $id)
     {
         $offer = Offer::findOrFail($id);
-        $offer->image = asset("storage/$offer->image");
+        $this->authorize('view', $offer);
         return response()->json($offer, 200);
     }
     /**
@@ -46,18 +45,20 @@ class OfferController extends Controller
      */
     public function create(OfferRequest $request)
     {
+        $this->authorize('create', Offer::class);
         try {
-            $supplier = Supplier::findOrFail($request->supplire_id);
+            $supplier = Supplier::findOrFail($request->supplier_id);
             $path = $request->file('image')->store('Offer', 'public');
             $Offer = Offer::create([
                 'supplier_id' => $supplier->id,
                 'image' => $path,
             ]);
-            /*$notification = new MobileNotificationServices;
-            $title = "عرض جديد";
-            $body = 'المورد' . $supplier->store_name . 'قام بإضافة عرض جديد';
-            $notification->sendNotificationToTopic('market', $title, $body);*/
-            $Offer->image = asset("storage/$path");
+            /** 
+             *$notification = new MobileNotificationServices;
+             *$title = "عرض جديد";
+             *$body = 'المورد' . $supplier->store_name . 'قام بإضافة عرض جديد';
+             *$notification->sendNotificationToTopic('market', $title, $body);
+             */
             return response()->json($Offer, 201);
         } catch (\Exception $e) {
             return response()->json($e->getMessage(), 500);
@@ -76,6 +77,7 @@ class OfferController extends Controller
         DB::beginTransaction();
         try {
             $offer = Offer::findOrFail($id);
+            $this->authorize('update', $offer);
 
             if ($request->hasFile('image')) {
                 $old_image = $offer->image;
@@ -96,9 +98,6 @@ class OfferController extends Controller
             ]);
 
             DB::commit();
-
-            $offer->image = asset("storage/$path");
-            $offer->image = asset("storage/$path");
             return response()->json($offer, 200);
         } catch (\Exception $e) {
             DB::rollback();
@@ -117,6 +116,7 @@ class OfferController extends Controller
         try {
             $offer = Offer::findOrFail($id);
             $old_image = $offer->image;
+            $this->authorize('delete', $offer);
 
             if (Storage::exists('public/' . $old_image)) {
                 Storage::delete('public/' . $old_image);
